requirements.txt
python-dotenv==1.0.0
SpeechRecognition==3.10.0
pyttsx3==2.90
openai==1.3.0
langchain==0.0.267
pywin32==306
PyQt5==5.15.9